moving data fetching to app.js
